+++ R/print.allPerms.R
@@ -2,8 +2,6 @@
 #' @noRd
 `print.allPerms` <- function(x, ...) {
     dims <- dim(x)
-    control <- attr(x, "control")
-    observed <- attr(x, "observed")
     attributes(x) <- NULL
     dim(x) <- dims
     print(x)