+++ R/print.how.R
@@ -36,19 +36,19 @@
                            prefix = pfix))
         writeLines(strwrap(paste("Permutation type:", ptype),
                            prefix = pfix))
-        if (isTRUE(all.equal(ptype, "partition"))) {
+        if (ptype == "partition") {
             sizes <- as.integer(table(plotStr))
             txt <- paste("Group sizes:", paste(sizes, collapse = ", "))
             writeLines(strwrap(txt, prefix = pfix))
             writeLines(strwrap("Within-group order retained", prefix = pfix))
         }
-        if (!isTRUE(all.equal(ptype, "partition"))) {
+        if (ptype != "partition") {
             mirrorP <- getMirror(x, which = "plots")
             writeLines(strwrap(paste("Mirrored?:",
                                      if(mirrorP) "Yes" else "No"),
                                prefix = pfix))
         }
-        if(isTRUE(all.equal(ptype, "grid"))) {
+        if(ptype == "grid") {
             symmetricP <- getSymmetric(x, which = "plots")
             writeLines(strwrap(paste("Symmetric?:",
                                      if(symmetricP) "Yes" else "No"),
@@ -78,7 +78,7 @@
         writeLines(strwrap(paste(txt, if(constantW) "No" else "Yes"),
                            prefix = pfix))
     }
-    if(isTRUE(all.equal(wtype, "grid"))) {
+    if(wtype == "grid") {
         symmetricW <- getSymmetric(x, which = "within")
         writeLines(strwrap(paste("Symmetric?:",
                                  if(symmetricW) "Yes" else "No"),