+++ R/update.how.R
@@ -31,9 +31,6 @@
 #' @export
 #' @noRd
 `update.how` <- function (object, ..., evaluate = TRUE) {
-    call <- getCall(object)
-    if (is.null(call))
-        stop("need an object with call component")
     extras <- match.call(expand.dots = FALSE)$...
 
     ## Preserve names when blocks or plots are not part of the update. If
@@ -43,21 +40,11 @@
     bname <- object$blocks.name
     pname <- object$plots$plots.name
 
-    ## process remaining ... args
-    if (length(extras)) {
-        existing <- !is.na(match(names(extras), names(call)))
-        ## do these individually to allow NULL to remove entries.
-        for (a in names(extras)[existing])
-            call[[a]] <- extras[[a]]
-        if (any(!existing)) {
-            call <- c(as.list(call), extras[!existing])
-            call <- as.call(call)
-        }
-    }
+    object.call <- amendCall(object, extras)
 
     ## probably want to evaluate hence default is TRUE
     if (evaluate) {
-        out <- eval(call, parent.frame())
+        out <- eval(object.call, parent.frame())
         ## add back in the chars we discovered earlier
         if (!update.blocks) {
             out$blocks.name <- bname
@@ -66,7 +53,7 @@
             out$plots$plots.name <- pname
         }
     } else {
-        out <- call
+        out <- object.call
     }
 
     out